+++ src/crates.rs
@@ -5,16 +5,19 @@ use std::io::Read;
 const ENDPOIONT: &str = "https://crates.io/api/v1/crates";
 
 pub fn url(crate_name: &str) -> String {
-  format!("{}/{}", ENDPOIONT, crate_name)
+    format!("{}/{}", ENDPOIONT, crate_name)
 }
 
 pub fn owners_url(crate_name: &str) -> String {
-  format!("{}/{}/owners", ENDPOIONT, crate_name)
+    format!("{}/{}/owners", ENDPOIONT, crate_name)
 }
 
 pub fn user_url(user_id: u64) -> String {
-  let per_page = 100;
-  format!("{}?page=1&per_page={}&sort=alpha&user_id={}", ENDPOIONT, per_page, user_id)
+    let per_page = 100;
+    format!(
+        "{}?page=1&per_page={}&sort=alpha&user_id={}",
+        ENDPOIONT, per_page, user_id
+    )
 }
 
 pub fn get(url: String) -> Option<serde_json::value::Value> {
@@ -22,18 +25,18 @@ pub fn get(url: String) -> Option<serde_json::value::Value> {
 
     match req {
         Err(e) => {
-          panic!("Error: {:?}", e);
-        },
+            panic!("Error: {:?}", e);
+        }
         Ok(mut res) => {
-          if res.status() == 200 {
-              let mut res_body = String::new(); 
-              res.read_to_string(&mut res_body)
-                  .expect("fail to read crate res_body");
-                  
-              Some(serde_json::from_str(&res_body).expect("fail to serde parse res_body"))
-          } else {
-            None 
-          }
-        },
+            if res.status() == 200 {
+                let mut res_body = String::new();
+                res.read_to_string(&mut res_body)
+                    .expect("fail to read crate res_body");
+
+                Some(serde_json::from_str(&res_body).expect("fail to serde parse res_body"))
+            } else {
+                None
+            }
+        }
     }
 }