+++ src/main.rs
@@ -3,8 +3,9 @@ extern crate reqwest;
 use colored::*;
 use std::env;
 
-mod format;
 mod crates;
+mod extract;
+mod format;
 
 fn main() {
     let args: Vec<String> = env::args().collect();
@@ -31,36 +32,50 @@ fn chit(crate_name: String) {
 
     match crates::get(crates::url(&crate_name)) {
         Some(crate_json) => {
-            // We found the crate
-            println!("{}", format::title_bar(width, &crate_name));
+            if let Some(fields) = extract::crate_fields(crate_json) {
+                println!("{}", format::title_bar(width, &crate_name));
 
-            let mut date = format::remove_quotes(crate_json["crate"]["updated_at"].to_string());
-            date.truncate(10);
-            // Version
-            padded_print(width, format!(
-                "Lastest version: {} ({})",
-                format::remove_quotes(crate_json["crate"]["max_version"].to_string()),
-                date
-            ));
+                let rating = extract::calculate_rating(fields.clone());
+                format::print_rating(width, rating);
 
-            // Download count
-            if let Some(download_count) = crate_json["crate"]["downloads"].as_i64() {
-                padded_print(width, format!("Download count: {:?}", download_count));
+                // Download count
+                if let Some(download_count) = fields.downloads {
+                    padded_print(width, format!("Total downloads: {}", download_count));
+                }
+
+                if let Some(recent_downloads) = fields.recent_downloads {
+                    padded_print(width, format!("Recent downloads: {}", recent_downloads));
+                }
+
+                padded_print(width, String::from("Versions:"));
+                for version in fields.versions {
+                    padded_print(
+                        width,
+                        format!(
+                            "    {} ({})",
+                            format::remove_quotes(version.semver),
+                            version.date
+                        ),
+                    );
+                }
             }
-        },
+        }
         None => {
             println!("{} {}", "Failed".red(), "to find that crate".magenta());
-            return
-        },
+            return;
+        }
     }
-    
+
     match crates::get(crates::owners_url(&crate_name)) {
         Some(crate_owners_json) => {
             // Owners
-            padded_print(width, format!(
-                "Owner: {}",
-                format::remove_quotes(crate_owners_json["users"][0]["name"].to_string())
-            ));
+            padded_print(
+                width,
+                format!(
+                    "Owner: {}",
+                    format::remove_quotes(crate_owners_json["users"][0]["name"].to_string())
+                ),
+            );
 
             padded_print(width, String::new());
             padded_print(width, String::from("Crates by owner:"));
@@ -70,18 +85,16 @@ fn chit(crate_name: String) {
                 if let Some(user_json) = crates::get(crates::user_url(user_id)) {
                     if let Some(array) = user_json["crates"].as_array() {
                         for thing in array {
-                            padded_print(width, format!(
-                                "    {}",
-                                format::remove_quotes(thing["id"].to_string())
-                            ));
+                            padded_print(
+                                width,
+                                format!("    {}", format::remove_quotes(thing["id"].to_string())),
+                            );
                         }
                     }
                 }
             }
-        },
-        None => {
-            println!("Failed to get crate owner details")
-        },
+        }
+        None => println!("Failed to get crate owner details"),
     }
 
     println!("{}", format::end_bar(width));